pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents[litellm]>=0.2.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
